delete create oke

+++ controllers/CompanyController.php
@@ -1,5 +1,18 @@
-<?php
+<!-- 
     // require_once __DIR__ . '/../views/pages/company.php';
+//pagination
+$pagination = mysqli_query($conn, "SELECT COUNT(companyID) AS total FROM company");
+$row = mysqli_fetch_assoc($pagination);
+$total_records = $row['total'];
+
+$current_page = isset($_GET['page']) ? $_GET['page'] : 1;
+$limit = 8;
+$total_page = ceil($total_records / $limit);
+
+if ($current_page > $total_page) $current_page = $total_page;
+        else if ($current_page < 1) $current_page = 1;
+
+$start = ($current_page - 1) * $limit >=0 ? ($current_page - 1) * $limit : 0;
 
 // read
 $dataCompanysSql = "SELECT companyID, code, name, description, email, phoneNumber, address FROM company";
@@ -33,6 +46,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
             document.getElementById('showData').hidden = false;";
     }
 }
+//read
 $dataClassesSql = "SELECT class.*, COUNT(student.studentID) AS numStudents
                   FROM class
                   LEFT JOIN student ON class.classID = student.classID
@@ -56,7 +70,53 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
         $scriptShowData = "
             document.getElementById('listClass').hidden = true;
             document.getElementById('showData').hidden = false;";
-    }}
+    }} -->
+
+
+    <?php
+    //pagination
+    $pagination = mysqli_query($conn, "SELECT COUNT(companyID) AS total FROM company");
+    $row = mysqli_fetch_assoc($pagination);
+    $total_records = $row['total'];
+    
+    $current_page = isset($_GET['page']) ? $_GET['page'] : 1;
+    $limit = 6;
+    $total_page = ceil($total_records / $limit);
+    
+    if ($current_page > $total_page) $current_page = $total_page;
+            else if ($current_page < 1) $current_page = 1;
+    
+    $start = ($current_page - 1) * $limit >=0 ? ($current_page - 1) * $limit : 0;
+    
+    // read
+    $dataCompanysSql = "SELECT company.*, COUNT(tour.tourID) AS numTours
+                      FROM company
+                      LEFT JOIN tour ON company.companyID = tour.companyID
+                      GROUP BY company.companyID
+                      LIMIT $start, $limit";
+    $dataCompanys = mysqli_query($conn, $dataCompanysSql);
+    $scriptShowData="";
+    if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+        if (isset($_POST['showData'])) {
+            $companyIDShow = $_POST["companyIDToShowTours"];
+            echo "<script>console.log('". $companyIDShow ."')</script>" ;
+            $dataCompanySql = "SELECT company.*, COUNT(tour.tourID) AS numTours
+                      FROM company
+                      LEFT JOIN tour ON company.companyID = tour.companyID
+                      WHERE company.companyID = $companyIDShow
+                      GROUP BY company.companyID";
+            $dataCompany = mysqli_query($conn, $dataCompanySql);    
+            $dataTourSql = "SELECT code, name, description, startDate, presentator, availables
+                               FROM tour
+                               WHERE companyID = $companyIDShow";
+            $dataTours = mysqli_query($conn, $dataTourSql);
+            $scriptShowData = "
+                document.getElementById('listCompany').hidden = true;
+                document.getElementById('showData').hidden = false;";
+        }
+    }
+
+
 // create
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     if (isset($_POST['createCompany'])) {
@@ -135,28 +195,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
                         }, 2000);
                     });
                 </script>";
-        }
-        // if (!mysqli_query($conn, $createCompanySql)) {
-        //     $errorMessage = mysqli_real_escape_string($conn, mysqli_error($conn));
-        //     echo "<script>
-        //             document.addEventListener('DOMContentLoaded', function() {
-        //                 document.getElementById('modalMessage').innerText = 'Sửa sinh viên thất bại: $errorMessage';
-        //                     $('#notificationModal').modal('show');
-        //                 });
-        //             </script>";
-            
-        // } else {
-        //     echo "<script>
-        //             document.addEventListener('DOMContentLoaded', function() {
-        //                  document.getElementById('modalMessage').innerText = 'Sửa sinh viên thành công';
-        //                 $('#notificationModal').modal('show');
-        //                 setTimeout(function(){
-        //                     window.location.href = '/PHP_Nhom3/index.php?controller=CompanyController';
-        //                 }, 2000);
-        //             });
-        //         </script>";
-                
-        // }
+        } 
     }
 }
 
@@ -164,32 +203,65 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     if (isset($_POST["deleteID"])) {
         $deleteId = $_POST["deleteID"];
-        $checkRoles = "SELECT * FROM company WHERE companyID = '$deleteId'";
-        $data = mysqli_query($conn, $checkRoles);
-        $row = mysqli_fetch_assoc($data);
-
-        $deleteSql = "DELETE FROM company WHERE companyID = '$deleteId'";
-        if (!mysqli_query($conn, $deleteSql)) {
+        
+        // Truy vấn để kiểm tra số lượng sinh viên của lớp học 
+        // Truy vấn để kiểm tra số lượng tour của doanh nghiệp
+        // $getQualityStudent = "SELECT COUNT(studentID) AS numStudents FROM student WHERE classID = '$deleteId'";
+        // $dataAll = mysqli_query($conn, $getQualityStudent);
+        $getQualityCompany = "SELECT COUNT(tourID) AS numTours FROM tour WHERE companyID = '$deleteId'";
+        $dataAll = mysqli_query($conn, $getQualityCompany);
+        if ($dataAll) {
+            $rowAll = mysqli_fetch_assoc($dataAll);
+            
+            if ($rowAll['numTours'] > 0) {
+                echo "<script>
+                        document.addEventListener('DOMContentLoaded', function() {
+                            document.getElementById('modalMessage').innerText = 'Không thể xóa doanh nghiệp này vì doanh nghiệp này vẫn còn chuyến tham quan.';
+                            $('#notificationModal').modal('show');
+                            setTimeout(function(){
+                                window.location.href = '/PHP_Nhom3/index.php?controller=CompanyController';
+                            }, 2000);
+                        });
+                    </script>";
+            } else {
+                // Xóa doanh nghiệp nếu không còn tour
+                $deleteSql = "DELETE FROM company WHERE companyID = '$deleteId'";
+                
+                if (mysqli_query($conn, $deleteSql)) {
+                    echo "<script>
+                            document.addEventListener('DOMContentLoaded', function() {
+                                document.getElementById('modalMessage').innerText = 'Doanh nghiệp đã được xóa thành công';
+                                $('#notificationModal').modal('show');
+                                setTimeout(function(){
+                                    window.location.href = '/PHP_Nhom3/index.php?controller=CompanyController';
+                                }, 2000);
+                            });
+                        </script>";
+                } else {
+                    $errorMessage = mysqli_real_escape_string($conn, mysqli_error($conn));
+                    echo "<script>
+                            document.addEventListener('DOMContentLoaded', function() {
+                                document.getElementById('modalMessage').innerText = 'Xóa doanh nghiệp thất bại: $errorMessage';
+                                $('#notificationModal').modal('show');
+                                setTimeout(function(){
+                                    window.location.href = '/PHP_Nhom3/index.php?controller=CompanyController';
+                                }, 2000);
+                            });
+                        </script>";
+                }
+            }
+        } else {
+            // Xử lý trường hợp có lỗi trong quá trình truy vấn
             $errorMessage = mysqli_real_escape_string($conn, mysqli_error($conn));
             echo "<script>
-                        document.addEventListener('DOMContentLoaded', function() {
-                            document.getElementById('modalMessage').innerText = 'Xóa doanh nghiệp thất bại: $errorMessage';
-                            $('#notificationModal').modal('show');
-                            setTimeout(function(){
-                                window.location.href = '/PHP_Nhom3/index.php?controller=CompanyController';
-                            }, 2000);
-                        });
-                      </script>";
-        } else {
-            echo "<script>
-                        document.addEventListener('DOMContentLoaded', function() {
-                            document.getElementById('modalMessage').innerText = 'Doanh nghiệp đã được xóa thành công';
-                            $('#notificationModal').modal('show');
-                            setTimeout(function(){
-                                window.location.href = '/PHP_Nhom3/index.php?controller=CompanyController';
-                            }, 2000);
-                        });
-                      </script>";
+                    document.addEventListener('DOMContentLoaded', function() {
+                        document.getElementById('modalMessage').innerText = 'Đã xảy ra lỗi khi kiểm tra số lượng tour: $errorMessage';
+                        $('#notificationModal').modal('show');
+                        setTimeout(function(){
+                            window.location.href = '/PHP_Nhom3/index.php?controller=CompanyController';
+                        }, 2000);
+                    });
+                </script>";
         }
     }
 }
